+++ src/lib.rs
@@ -1,6 +1,6 @@
 pub mod cipher_suite;
 pub mod commands;
 pub mod error;
-pub mod file_ops;
+pub mod header;
 pub mod persona;
 pub mod wallet;
